+++ CHANGELOG.md
@@ -6,6 +6,8 @@
     or `@Dart(Skip)`). Elements marked with this attribute will be omitted in that platform's
     generated code. This attribute is not supported for C++.
   * Added support for creating an interface implementation directly from a set of lambdas in Dart.
+  * Added `-werror` command line parameter to support elevating specific validation warnings to
+    errors. Current supported warning types are `DocLinks` and `DartOverloads`.
 
 ## 6.5.0
 Release date: 2020-04-16